搜索功能实现：记录算法时间、搜索一次返回搜素结果的同时会在数据库表中添加一行搜索记录 待完成的接口，用户评分

+++ src/main/java/csu/krystal/recommsys/entity/Record.java
@@ -30,7 +30,7 @@ public class Record implements Serializable {
 
     private String content;
 
-    private Float runTime;
+    private double runTime;
 
     @JsonFormat(pattern = "yyyy-MM-dd", timezone = "GMT+8")
     private Date createTime;
